Revert "Admin front back int"

+++ backend/login-microservice/src/main/java/edu/oswego/cs/database/ProfessorCheck.java
@@ -31,7 +31,6 @@ public class ProfessorCheck {
         }
     }
 
-    @Deprecated(since = "1.0.2", forRemoval = false)
     public void addProfessors() throws IOException {
         String path = getPath();
         var reader = new BufferedReader(new FileReader(path + "professor-list.txt"));
@@ -58,18 +57,14 @@ public class ProfessorCheck {
 
         for (Document professor : professorCollection.find()) {
             String professorID = professor.getString("professor_id");
-            if (!professorList.contains(professorID))
-                professorCollection.deleteOne(eq("professor_id", professorID));
+            if (!professorList.contains(professorID)) professorCollection.deleteOne(eq("professor_id", professorID));
         }
 
         reader.close();
     }
 
-    @Deprecated(since = "1.0.2", forRemoval = false)
     public String getPath() {
-        String path = (System.getProperty("user.dir").contains("\\"))
-                ? System.getProperty("user.dir").replace("\\", "/")
-                : System.getProperty("user.dir");
+        String path = (System.getProperty("user.dir").contains("\\")) ? System.getProperty("user.dir").replace("\\", "/") : System.getProperty("user.dir");
         String[] slicedPath = path.split("/");
         String targetDir = "defaultServer";
         StringBuilder relativePathPrefix = new StringBuilder();
@@ -85,5 +80,4 @@ public class ProfessorCheck {
 
         return relativePathPrefix.toString();
     }
-
 }
